base-requirements.txt
Flask==0.10.1
Flask-Cache==0.13.1
Flask-Script==2.0.5
Flask-Cors==3.0.2
loremipsum==1.0.5
python-amazon-simple-product-api==1.5.0
python-slugify==0.0.9
requests==2.13.0
PyYAML==3.12
pkutils>=0.13.6,<0.14.0
docutils==0.13.1
sphinxcontrib-napoleon==0.6.1
